pyproject.toml
[project]
name = "novelcraft"
version = "0.1.0"
description = "이지연 판타지 소설 자동 집필 시스템"
requires-python = ">=3.11"
dependencies = [
    "anthropic>=0.40.0",
    "click>=8.1.0",
    "python-dotenv>=1.0.0",
    "rich>=13.7.0",
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "google-genai>=1.0.0",
    "python-docx>=1.0.0",
]

[project.scripts]
novelcraft = "novelcraft.cli:cli"

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["novelcraft*"]
